+++ dao/src/main/java/com/eghm/dao/mapper/TaskConfigMapper.java
@@ -15,7 +15,7 @@ public interface TaskConfigMapper {
      *
      * @param id 条件 
      */
-    TaskConfig selectByPrimaryKey(Integer id);
+    TaskConfig selectByPrimaryKey(Long id);
 
     /**
      * 根据主键来更新部分数据库记录
