+++ model/src/main/java/com/eghm/model/dto/address/AddressEditDTO.java
@@ -33,14 +33,14 @@ public class AddressEditDTO {
      */
     @ApiModelProperty(value = "省份id",required = true)
     @NotNull(message = "省份不能为空")
-    private Integer provinceId;
+    private Long provinceId;
 
     /**
      * 城市id
      */
     @ApiModelProperty(value = "城市id",required = true)
     @NotNull(message = "城市不能为空")
-    private Integer cityId;
+    private Long cityId;
 
     /**
      * 县区id
