+++ service/src/main/java/com/eghm/service/common/impl/BannerServiceImpl.java
@@ -40,7 +40,7 @@ public class BannerServiceImpl implements BannerService {
     }
 
     @Override
-    public Banner getById(Integer id) {
+    public Banner getById(Long id) {
         return bannerMapper.selectByPrimaryKey(id);
     }
