+++ service/src/main/java/com/eghm/service/sys/impl/SysDeptServiceImpl.java
@@ -44,7 +44,7 @@ public class SysDeptServiceImpl implements SysDeptService {
     }
 
     @Override
-    public SysDept getById(Integer id) {
+    public SysDept getById(Long id) {
         return sysDeptMapper.selectByPrimaryKey(id);
     }
